Create Dice Number Generator In C
This project is a simple C program that generates a random number between 1 and 6, simulating the roll of a dice. Each time the program is run, it produces a different result, thanks to the use of the current time as a seed for the random number generator.

Features
Random Number Generation: Utilizes the rand() function to generate a pseudo-random number.
Range Control: Ensures the number is between 1 and 6, mimicking a standard dice roll.
Seeding with Current Time: Uses srand(time(0)) to seed the random number generator, providing different outcomes for each execution.